refactor: pageコンポーネントを今のやり方に変更、メニュー状態をコンテクストに分離(#13)
