+++ src/components/common/Layout/Layout.tsx
@@ -1,5 +1,4 @@
 import { FC } from 'react';
-import Head from 'next/head';
 import { css } from '@emotion/react';
 import Footer from '@/components/common/Footer';
 
@@ -18,13 +17,6 @@ const Layout: FC<Props> = ({ inertFlg, children }) => {
           : node.removeAttribute('inert'))
       }
     >
-      <Head>
-        <title>Windbnb</title>
-        <meta
-          name="description"
-          content="devChallenges.io - Windbnb | by h-yoshikawa44"
-        />
-      </Head>
       <div css={contents}>{children}</div>
       <Footer css={customFooter} />
     </div>
