+++ src/components/context/DrawerMenuContext/index.ts
@@ -1,0 +1,3 @@
+import { DrawerMenuContext, DrawerMenuProvider } from './DrawerMenuContext';
+
+export { DrawerMenuContext, DrawerMenuProvider };
