+++ src/hooks/common/index.ts
@@ -1,0 +1,3 @@
+import useDrawerMenu from './useDrawerMenu';
+
+export { useDrawerMenu };
