+++ src/hooks/common/useDrawerMenu/index.ts
@@ -1,0 +1,3 @@
+import useDrawerMenu from './useDrawerMenu';
+
+export default useDrawerMenu;
